+++ PoCFinal/Modules/ContactRequests/Domain/ContactRequestStatus.cs
@@ -1,6 +1,0 @@
-namespace ContactRequests.Domain;
-
-public enum ContactRequestStatus
-{
-    Pending = 0
-}